Improved filelocking in convert.sh Added a bug-reporting section to process-upload added touch ./incoming/DONE/index.html to build.xml
git-svn-id: https://xmlresume.svn.sourceforge.net/svnroot/xmlresume@521 58bf4eea-5411-0410-84c9-993ea22221af

+++ trunk/resume-website/src/www/orc/convert.sh
@@ -1,4 +1,4 @@
-#!/bin/sh
+#!/bin/bash
 # convert.sh
 # Contributed 2002 by Mark Miller (brandondoyle)
 #
@@ -36,22 +36,22 @@ for jarfile in `ls -1 ${SUPPORT_HOME}/ant/lib | grep .jar`; do
 done
 export CLASSPATH=$cp
 
-echo "Using ClassPath: $CLASSPATH"
-
 # Add option for the CLASSPATH
 ANT_OPTS="${ANT_OPTS} -classpath ${LOCALCLASSPATH}"
 
 cd @WWW_ROOT_FS@/orc/incoming
 for resume in `ls -1 | grep '^orc'`; 
 do
-	# If the lockfile does not exist, then 
-	if [ ! -e .$resume ]; then
-		cd $resume
-		grep '^email =' user.props | cut -f 3 -d" " | ${MD5CMD} >> ../../users.md5
-		${ANTCMD} -verbose -propertyfile user.props \
-		-find build.xml dispatch >& ./out/antlog.txt
-#		sendmail -f'noreply@xmlresume.sourceforge.net' -t < reply.email
- 		cd ..
-  		mv $resume DONE/$resume
-	fi
+   # If no lockfiles exist, create the .convert lockfile and convert
+   if [ ! -e .UPLOAD.$resume -a ! -e .CONVERT.$resume ]; then
+      touch .convert.$resume
+	cd $resume
+	grep '^email =' user.props | cut -f 3 -d" " | ${MD5CMD} >> ../../users.md5
+	${ANTCMD} -verbose -propertyfile user.props \
+	-find build.xml dispatch >& ./out/antlog.txt
+#	sendmail -f'noreply@xmlresume.sourceforge.net' -t < reply.email
+	cd ..
+	mv $resume DONE/$resume
+      rm -f .CONVERT.$resume
+    fi
 done
